bootstrap: truncate config file when saving
Before this change config files would have remnants of the previous
file's data. When saving, open file using os.Create() to truncate and/or create
the file before writing. In cases when the file is read only, open using
os.Open() which is shorthand for opening with the os.O_RDWR flag.