Add the option to lock factorio server port to the defualt game port.
Add a conf.json option "factorio_port_lock". When set to true, force
the game to run on the default "factorio_port" (--game-port on command
line flag). This prevents confused admins when they run the game on
a different port and can't get a connection. It also grays out the
port field so it is not editable on the FSM Controls tab.